feat 69 mock values inversed

+++ app/src/main/java/edu/iesam/meceiot/features/pantallasensor/data/local/db/SensorEntity.kt
@@ -12,8 +12,8 @@ class SensorEntity(
     @PrimaryKey @ColumnInfo(name = "id") val id: Int,
     @ColumnInfo(name = "sensorName") val nombre: String,
     @ColumnInfo(name = "panelName") val nombrePanel: String,
-    @ColumnInfo(name = "values_x") val valoresX: List<Int>,
-    @ColumnInfo(name = "values_y") val valoresY: List<Long>,
+    @ColumnInfo(name = "values_x") val valoresX: List<Long>,
+    @ColumnInfo(name = "values_y") val valoresY: List<Int>,
     @ColumnInfo(name = "leyend_x") val leyendaX: String,
     @ColumnInfo(name = "leyend_y") val leyendaY: String,
     @ColumnInfo(name = "createdAt") val date: Date
